requirements.txt
openai>=1.0.0
anthropic>=0.21.0
google-generativeai>=0.8.0